+++ docs/api/tasks/keyword_extraction.md
@@ -1,6 +1,6 @@
 # Keyword Extraction Task
 
-::: openaivec.task.tasks.keyword_extraction
+::: openaivec.task.nlp.keyword_extraction
     options:
       show_source: true
       show_root_heading: true
